jam domain been cleaned up
